+++ desktop/src/main/lib/code-editor.ts
@@ -1,0 +1,3 @@
+export const handleCodeEditor = (_, payload) => {
+  console.log(payload);
+};