+++ desktop/src/renderer/src/constants/base.ts
@@ -1,0 +1,9 @@
+export const baseUri = "https://145.223.97.55:9292";
+
+// export const baseUri =
+//   // @ts-ignore
+//   import.meta.env.MODE !== "development"
+//     ? // @ts-ignore
+//       import.meta.env.RENDERER_VITE_SERVER_BASE_URL
+//     : // @ts-ignore
+//       import.meta.env.RENDERER_VITE_SERVER_BASE_URL_LOCAL;