+++ changelog.md
@@ -31,3 +31,8 @@
 - [chore] Created executeme platform logo for free
 - [feat] Created a release notes ci/cd pipeline to automate our release version
 - [feat] Create electron builder CI for build various platform applicaiton
+- [feat] ShadcnUI added to our desktop application
+- [feat] Monaco editor setup with electron js
+- [feat] Code highlight & inte. is supported
+- [fix] electron main process api calling issues
+- [feat] api calling bridge for secure our application